Fixed tooltip positioning for top elements
